WIP: Attempting to NOT wipe high scores when reseting the game. This requires storing the highscores in stack memory, then clearing all RAM but stack memory, then restoring scores from stack memory, and THEN clearing stack memory.
This all works, accept that it doesn't handle random RAM values, resulting in crazy scores on first boot. To solve this, I need to write some sort of pre-determined sequence to RAM in main(), and then check that on restart. If it doesn't match, assume this is a fresh boot!